Fixed steup files

+++ shells/usersetup.sh
@@ -49,8 +49,13 @@ cloneDotfiles() {
 
     cd ~/tools
 
-    # Clone dotfiles
-    git clone https://github.com/Yoswell/dotfiles.git ~/tools/dotfiles
+    # Clone dotfiles (si ya existe, actualizar)
+    if [ -d ~/tools/dotfiles ]; then
+        cd ~/tools/dotfiles
+        git pull
+    else
+        git clone https://github.com/Yoswell/dotfiles.git ~/tools/dotfiles
+    fi
 }
 
 # Setting up wallpapers
@@ -62,9 +67,13 @@ setupWallpapers() {
     # Copy wallpapers from dotfiles
     cp -rf wallpapers ~/Desktop/
     
-    # Set wallpapers
-    sudo cp -f wallpapers/xcyberpunk.png /usr/share/backgrounds/kali/kali-maze-16x9.jpg
-    sudo cp -f wallpapers/xcyberpunk.png /usr/share/backgrounds/kali/login.svg
+    # Set wallpapers (verificar si existen los directorios)
+    if [ -d /usr/share/backgrounds/kali ]; then
+        sudo cp -f wallpapers/xcyberpunk.png /usr/share/backgrounds/kali/kali-maze-16x9.jpg
+        sudo cp -f wallpapers/xcyberpunk.png /usr/share/backgrounds/kali/login.svg
+    else
+        warning "Kali backgrounds directory not found, skipping wallpaper setup"
+    fi
 }
 
 # Creating directories
@@ -81,7 +90,6 @@ setupBinAndZsh() {
 
     # Oh My Zsh
     rm -rf ~/.oh-my-zsh
-    rm -rf powerlevel10k
     sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
     
     # Create bin directory if it doesn't exist
@@ -92,18 +100,23 @@ setupBinAndZsh() {
 
     # ZSH and BIN of normal user
     cp -rf zshrc ~/.zshrc
-    cp -rf bin ~/ && chmod +x ~/bin/*
+    cp -rf bin ~/ && chmod +x ~/bin/* 2>/dev/null
     
-    # ZSH plugins
-    git clone https://github.com/zsh-users/zsh-autosuggestions.git ~/.oh-my-zsh/custom/plugins
-    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins
-    git clone https://github.com/Yoswell/zsh_tshark_autocomplete.git ~/.oh-my-zsh/custom/plugins
+    # ZSH plugins (limpiar antes de clonar)
+    rm -rf ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
+    rm -rf ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
+    rm -rf ~/.oh-my-zsh/custom/plugins/zsh_tshark_autocomplete
+    
+    git clone https://github.com/zsh-users/zsh-autosuggestions.git ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
+    git clone https://github.com/Yoswell/zsh_tshark_autocomplete.git ~/.oh-my-zsh/custom/plugins/zsh_tshark_autocomplete
     
     # Powerlevel10k
-    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ~/.oh-my-zsh
+    rm -rf ~/.oh-my-zsh/custom/themes/powerlevel10k
+    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ~/.oh-my-zsh/custom/themes/powerlevel10k
     
-    # ZSH and BIN of root user (NOT HERE - moved to setup-root.sh)
-    # sudo cp -f zshrc /root/.zshrc  # <-- MOVED TO ROOT SCRIPT
+    # Configurar tema en .zshrc
+    sed -i 's|ZSH_THEME=.*|ZSH_THEME="powerlevel10k/powerlevel10k"|' ~/.zshrc
 }
 
 # Setting up user themes
@@ -120,7 +133,7 @@ setupUserThemes() {
     mkdir -p .config/xfce4/xfconf/xfce-perchannel-xml
     mkdir -p .config/gtk-3.0
     
-    cd ~/tools/dotfiles
+    cd ~/tools/dotfiles/config  # ← CAMBIADO: entrar a config/
     
     # Extract themes from themes directory
     tar -xf themes/mantinight.tar -C ~/.themes/
@@ -131,35 +144,40 @@ setupUserThemes() {
     cp -f schemes/kitty.conf ~/.config/kitty/
     cp -f schemes/tmux.conf ~/.tmux.conf
     
-    # Copy polybar configuration
-    cp -f polybar/config.ini ~/.config/polybar/
-    cp -rf polybar/scripts ~/.config/polybar/
-    chmod +x ~/.config/polybar/scripts/*.sh
+    # Copy polybar configuration (ahora dentro de config/)
+    if [ -f polybar/config.ini ]; then
+        cp -f polybar/config.ini ~/.config/polybar/
+        cp -rf polybar/scripts ~/.config/polybar/
+        chmod +x ~/.config/polybar/scripts/*.sh 2>/dev/null
+    fi
 
     # Copy rofi theme system wide
     sudo cp -f schemes/rofi_theme.rasi /usr/share/rofi/themes/
     
-    # Update XFCE configs
-    cp -f xfce/xfce4-keyboard-shortcuts.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/
-    cp -f xfce/xfce4-power-manager.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/
-    cp -f xfce/xfce4Settings.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/
-    cp -f xfce/xfce4-desktop.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/
-    cp -f xfce/thunar.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/
+    # Update XFCE configs (están en la raíz de dotfiles)
+    cd ~/tools/dotfiles  # ← Volver a la raíz para XFCE
+    if [ -d xfce ]; then
+        cp -f xfce/xfce4-keyboard-shortcuts.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/ 2>/dev/null
+        cp -f xfce/xfce4-power-manager.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/ 2>/dev/null
+        cp -f xfce/xfce4Settings.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/ 2>/dev/null
+        cp -f xfce/xfce4-desktop.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/ 2>/dev/null
+        cp -f xfce/thunar.xml ~/.config/xfce4/xfconf/xfce-perchannel-xml/ 2>/dev/null
+    fi
 }
 
 # Installing valuable fonts
 installFonts() {
     section "Installing fonts"
 
-    cd ~/tools/dotfiles/fonts
-
-    # Copy fonts to desktop for backup
-    cp -rf jetbrains.zip ~/Desktop/jetbrains
-    cp -rf protonerd.zip ~/Desktop/protonerd
+    cd ~/tools/dotfiles/config/fonts  # ← CAMBIADO: fonts está dentro de config/
+    
+    # Crear directorios de fuentes
+    sudo mkdir -p /usr/share/fonts/truetype/jetbrains-mono
+    sudo mkdir -p /usr/share/fonts/truetype/proto-nerd
 
     # Install fonts
-    sudo unzip jetbrains.zip -d /usr/share/fonts/truetype/jetbrains-mono
-    sudo unzip protonerd.zip -d /usr/share/fonts/truetype/proto-nerd
+    sudo unzip -o jetbrains.zip -d /usr/share/fonts/truetype/jetbrains-mono
+    sudo unzip -o protonerd.zip -d /usr/share/fonts/truetype/proto-nerd
     
     sudo fc-cache -f -v
 }
@@ -174,6 +192,8 @@ runToolsScript() {
     if [ -f shells/tools.sh ]; then
         chmod +x shells/tools.sh
         ./shells/tools.sh
+    else
+        warning "tools.sh not found in shells/ directory"
     fi
 }
 
@@ -190,8 +210,8 @@ main() {
     
     # Installation sequence
     uninstallPackages
-    runToolsScript
     cloneDotfiles
+    runToolsScript
     setupWallpapers
     createDirectories
     setupBinAndZsh
